refactor: separate AlignDescriptors and RestoreDescriptorsPosition into individual files
