+++ packages/ketcher-core/src/application/editor/operations/descriptors.ts
@@ -13,67 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  ***************************************************************************/
-/* eslint-disable @typescript-eslint/no-use-before-define */
-import { BaseOperation } from './BaseOperation';
-import { OperationType } from './OperationType';
-import { ReStruct } from '../../render';
-import { Vec2 } from 'domain/entities/vec2';
 
-// todo: separate classes: now here is circular dependency in `invert` method
+import { AlignDescriptors } from './AlignDescriptors';
+import { RestoreDescriptorsPosition } from './RestoreDescriptorsPosition';
 
-class AlignDescriptors extends BaseOperation {
-  readonly history: any;
-
-  constructor() {
-    super(OperationType.ALIGN_DESCRIPTORS);
-    this.history = {};
-  }
-
-  execute(restruct: ReStruct) {
-    const struct = restruct.molecule;
-    const sgroups: any[] = Array.from(struct.sgroups.values()).reverse();
-
-    const structBox: any = struct.getCoordBoundingBoxObj();
-    let alignPoint = new Vec2(structBox.max.x, structBox.min.y).add(
-      new Vec2(2.0, -1.0),
-    );
-
-    sgroups.forEach((sgroup) => {
-      this.history[sgroup.id] = new Vec2(sgroup.pp);
-      alignPoint = alignPoint.add(new Vec2(0.0, 0.5));
-      sgroup.pp = alignPoint;
-      struct.sgroups.set(sgroup.id, sgroup);
-      BaseOperation.invalidateItem(restruct, 'sgroupData', sgroup.id, 1);
-    });
-  }
-
-  invert() {
-    return new RestoreDescriptorsPosition(this.history);
-  }
-}
-
-class RestoreDescriptorsPosition extends BaseOperation {
-  readonly history: any;
-
-  constructor(history: any) {
-    super(OperationType.RESTORE_DESCRIPTORS_POSITION);
-    this.history = history;
-  }
-
-  execute(restruct: ReStruct) {
-    const struct = restruct.molecule;
-    const sgroups: any[] = Array.from(struct.sgroups.values());
-
-    sgroups.forEach((sgroup) => {
-      sgroup.pp = this.history[sgroup.id];
-      struct.sgroups.set(sgroup.id, sgroup);
-      BaseOperation.invalidateItem(restruct, 'sgroupData', sgroup.id, 1);
-    });
-  }
-
-  invert() {
-    return new AlignDescriptors();
-  }
-}
+AlignDescriptors.InverseConstructor = RestoreDescriptorsPosition;
+RestoreDescriptorsPosition.InverseConstructor = AlignDescriptors;
 
 export { AlignDescriptors, RestoreDescriptorsPosition };
